moved LuceneSearchUtil to LuceneUtils
