+++ src/main/java/com/celements/search/lucene/LuceneUtils.java
@@ -1,17 +1,32 @@
 package com.celements.search.lucene;
 
+import static com.google.common.base.MoreObjects.*;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 import org.python.google.common.base.Strings;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.xwiki.model.EntityType;
 import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.EntityReference;
 
 import com.celements.model.util.ModelUtils;
+import com.celements.search.lucene.query.IQueryRestriction;
+import com.celements.search.lucene.query.QueryRestrictionGroup;
+import com.celements.search.lucene.query.QueryRestrictionGroup.Type;
 import com.google.common.base.Function;
+import com.google.common.base.Splitter;
 import com.xpn.xwiki.web.Utils;
 
 public class LuceneUtils {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(LuceneUtils.class);
+
   private static final String QUOTE = "\"";
+  private static final Splitter SPLITTER = Splitter.on("[,;\\| ]+");
 
   public static final Function<String, String> FUNC_EXACTIFY = new Function<String, String>() {
 
@@ -32,6 +47,14 @@ public class LuceneUtils {
       }
     }
     return str;
+  }
+
+  public static List<String> exactify(List<String> strs) {
+    List<String> ret = new ArrayList<>();
+    for (String str : firstNonNull(strs, Collections.<String>emptyList())) {
+      ret.add(exactify(str));
+    }
+    return ret;
   }
 
   public static String asFieldName(DocumentReference classRef, String field) {
@@ -68,6 +91,26 @@ public class LuceneUtils {
     } else {
       return "";
     }
+  }
+
+  public static IQueryRestriction buildRestrictionGroup(Type type, String valuesStr,
+      Function<String, IQueryRestriction> restrictionFunc) {
+    return buildRestrictionGroup(type, SPLITTER.split(valuesStr), restrictionFunc);
+  }
+
+  public static IQueryRestriction buildRestrictionGroup(Type type, Iterable<String> values,
+      Function<String, IQueryRestriction> restrictionFunc) {
+    QueryRestrictionGroup grp = new QueryRestrictionGroup(type);
+    for (String str : firstNonNull(values, Collections.<String>emptyList())) {
+      if (Strings.nullToEmpty(str).trim().isEmpty()) {
+        try {
+          grp.add(restrictionFunc.apply(str));
+        } catch (IllegalArgumentException iae) {
+          LOGGER.warn("building restriction failed for value '{}' ", str);
+        }
+      }
+    }
+    return grp;
   }
 
   private static ModelUtils getModelUtils() {
